separate command vs helper
