+++ src/commands/rule34/rule34.interface.ts
@@ -20,8 +20,8 @@ export interface Rule34KeywordList {
 }
 
 export interface Rule34CommandKeyList {
-  RULE_34_DELETE_KEYWORD: "~rule34deleteKeyword";
-  RULE_34_ADD_KEYWORD: "~rule34addKeyword";
+  RULE34_DELETE_KEYWORD: "~rule34deleteKeyword";
+  RULE34_ADD_KEYWORD: "~rule34addKeyword";
   RULE34_SEARCH: "~rule34";
   RULE34_SET_RECURRING: "~rule34setRecurring";
   RULE34_GET_RECURRING: "~rule34getRecurring";
@@ -30,8 +30,8 @@ export interface Rule34CommandKeyList {
 }
 
 export interface Rule34CommandList {
-  RULE_34_DELETE_KEYWORD: Command;
-  RULE_34_ADD_KEYWORD: Command;
+  RULE34_DELETE_KEYWORD: Command;
+  RULE34_ADD_KEYWORD: Command;
   RULE34_SEARCH: Command;
   RULE34_SET_RECURRING: Command;
   RULE34_GET_RECURRING: Command;
